added create inventory script in build.sh
