+++ scenarios/scenario1/resources/vm.tf
@@ -17,6 +17,7 @@ resource "opentelekomcloud_compute_instance_v2" "http" {
   network {
     port             = opentelekomcloud_networking_port_v2.http.*.id[count.index]
   }
+  tag                = { "group": "gatewayed" }
 }
 
 # Create network port
